quiet unsigned long compare.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V4-2-patches@5046 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/agentx/master_request.c
@@ -509,7 +509,7 @@ agentx_add_request( struct agent_snmp_session *asp,
 			   vbp->name, vbp->name_length, ASN_PRIV_EXCL_RANGE,
 			   (u_char*)sub->end, sub->end_len*sizeof(oid));
     }
-    if ( sub->timeout > request->pdu->time ) {
+    if ( sub->timeout > (int)request->pdu->time ) {
 	request->pdu->time = sub->timeout;
 	request->pdu->flags |= UCD_MSG_FLAG_PDU_TIMEOUT;
     }
